Move attributes to cmd As the attributes are just a structured input of the CLI I think they fit better there
