+++ cmd/configure/attributes/container/container.go
