+++ cmd/configure/attributes/container/image.go
