+++ cmd/configure/attributes/container/image_test.go
