+++ cmd/configure/attributes/pod/pod.go
