+++ cmd/configure/attributes/pod/pod_test.go
