+++ pkg/configure/enrichment/metadata/content.go
@@ -5,8 +5,8 @@ import (
 	"maps"
 	"strings"
 
-	"github.com/Dynatrace/dynatrace-bootstrapper/pkg/configure/attributes/container"
-	"github.com/Dynatrace/dynatrace-bootstrapper/pkg/configure/attributes/pod"
+	"github.com/Dynatrace/dynatrace-bootstrapper/cmd/configure/attributes/container"
+	"github.com/Dynatrace/dynatrace-bootstrapper/cmd/configure/attributes/pod"
 	"github.com/Dynatrace/dynatrace-bootstrapper/pkg/utils/structs"
 	"github.com/pkg/errors"
 )
